requirements.txt
arts==1
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
docutils==0.16
feedparser==5.2.1
idna==2.10
Kivy-Garden==0.1.4
meds==1215
numpy==1.19.4
pandas==1.1.4
PasteDeploy==2.1.1
Pillow==8.0.1
playsound==1.2.2
prompt-toolkit==1.0.14
psutil==5.7.3
psycopg2-binary==2.8.6
pulse==0.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.0.0
Pygments==2.7.2
PyInquirer==1.0.3
python-dateutil==2.8.1
pytz==2020.4
regex==2020.11.13
requests==2.25.0
six==1.15.0
sleekxmpp==1.3.3
soupsieve==2.0.1
urllib3==1.26.2
wcwidth==0.2.5
WebOb==1.8.6
